+++ covalent/__init__.py
@@ -42,6 +42,7 @@ from ._shared_files.util_classes import RESULT_STATUS as status  # nopycln: impo
 from ._workflow import (  # nopycln: import
     DepsBash,
     DepsCall,
+    DepsModule,
     DepsPip,
     Lepton,
     TransportableObject,